refactor: kafka event 사용 모듈로 위치 이동 및 미사용 event 정리 - (#349)
