--- common/src/main/java/com/back/common/user/event/WalletCreateRequestedEvent.java
@@ -1,8 +1,0 @@
-package com.back.common.user.event;
-
-import com.back.common.event.EventName;
-
-public record WalletCreateRequestedEvent(
-         Long userId
-) implements EventName {
-}
